requirements.txt
streamlit
pandas
matplotlib
scikit-learn
pyyaml
